LPS-76234 Reorganize fragments editor components
