+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/edit_layout_page_template_entry.jsp
@@ -85,6 +85,6 @@ layoutPageTemplateEditorContext.put("updatePageTemplateURL", String.valueOf(edit
 
 <soy:template-renderer
 	context="<%= layoutPageTemplateEditorContext %>"
-	module="layout-admin-web/js/fragments_editor/LayoutPageTemplateEditor.es"
-	templateNamespace="LayoutPageTemplateEditor.render"
+	module="layout-admin-web/js/fragments_editor/FragmentsEditor.es"
+	templateNamespace="FragmentsEditor.render"
 />
