+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/Fragment.es.js
@@ -4,13 +4,13 @@ import Component from 'metal-component';
 import {Config} from 'metal-state';
 import Soy from 'metal-soy';
 
-import templates from './LayoutPageTemplateFragment.soy';
+import templates from './Fragment.soy';
 
 /**
- * LayoutPageTemplateFragment
+ * Fragment
  * @review
  */
-class LayoutPageTemplateFragment extends Component {
+class Fragment extends Component {
 	/**
 	 * @inheritDoc
 	 * @review
@@ -220,13 +220,13 @@ class LayoutPageTemplateFragment extends Component {
  * @static
  * @type {!Object}
  */
-LayoutPageTemplateFragment.STATE = {
+Fragment.STATE = {
 	/**
 	 * Editable values that should be used instead of the default ones
 	 * inside editable fields.
 	 * @default {}
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {!Object}
 	 */
@@ -236,7 +236,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Fragment entry ID
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {!string}
 	 */
@@ -246,7 +246,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Fragment index
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {!number}
 	 */
@@ -256,7 +256,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Fragment name
 	 * @default ''
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {string}
 	 */
@@ -266,7 +266,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Portlet namespace needed for prefixing form inputs
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {!string}
 	 */
@@ -276,7 +276,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * URL for getting a fragment render result.
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {!string}
 	 */
@@ -286,7 +286,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Fragment spritemap
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @review
 	 * @type {!string}
 	 */
@@ -296,7 +296,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Fragment content to be rendered
 	 * @default Soy.toIncDom('')
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @private
 	 * @review
 	 * @type {function}
@@ -309,7 +309,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * List of AlloyEditor instances used for inline edition
 	 * @default []
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @private
 	 * @review
 	 * @type {Array<AlloyEditor>}
@@ -322,7 +322,7 @@ LayoutPageTemplateFragment.STATE = {
 	 * Flag indicating that fragment information is being loaded
 	 * @default false
 	 * @instance
-	 * @memberOf LayoutPageTemplateFragment
+	 * @memberOf Fragment
 	 * @private
 	 * @review
 	 * @type {boolean}
@@ -330,7 +330,7 @@ LayoutPageTemplateFragment.STATE = {
 	_loading: Config.bool().value(false),
 };
 
-Soy.register(LayoutPageTemplateFragment, templates);
+Soy.register(Fragment, templates);
 
-export {LayoutPageTemplateFragment};
-export default LayoutPageTemplateFragment;
+export {Fragment};
+export default Fragment;
