+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/Fragment.soy
@@ -1,7 +1,7 @@
-{namespace LayoutPageTemplateFragment}
+{namespace Fragment}
 
 /**
- * LayoutPageTemplateFragment
+ * Fragment
  */
 {template .render}
 	{@param spritemap: string}
