+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/sidebar/SidebarAddedFragment.es.js
@@ -2,12 +2,12 @@ import Component from 'metal-component';
 import {Config} from 'metal-state';
 import Soy from 'metal-soy';
 
-import templates from './LayoutPageTemplateSidebarAddedFragment.soy';
+import templates from './SidebarAddedFragment.soy';
 
 /**
- * LayoutPageTemplateSidebarAddedFragment
+ * SidebarAddedFragment
  */
-class LayoutPageTemplateSidebarAddedFragment extends Component {
+class SidebarAddedFragment extends Component {
 	/**
 	 * Callback executed when the fragment remove button is clicked.
 	 * It emits a 'fragmentRemoveButtonClick' event with the fragment index.
@@ -26,12 +26,12 @@ class LayoutPageTemplateSidebarAddedFragment extends Component {
  * @static
  * @type {!Object}
  */
-LayoutPageTemplateSidebarAddedFragment.STATE = {
+SidebarAddedFragment.STATE = {
 	/**
 	 * Fragment index
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateSidebarAddedFragment
+	 * @memberOf SidebarAddedFragment
 	 * @type {!number}
 	 */
 	index: Config.number().required(),
@@ -40,7 +40,7 @@ LayoutPageTemplateSidebarAddedFragment.STATE = {
 	 * Fragment name
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateSidebarAddedFragment
+	 * @memberOf SidebarAddedFragment
 	 * @type {!string}
 	 */
 	name: Config.string().required(),
@@ -49,13 +49,13 @@ LayoutPageTemplateSidebarAddedFragment.STATE = {
 	 * Path of the available icons.
 	 * @default undefined
 	 * @instance
-	 * @memberOf LayoutPageTemplateSidebarAddedFragment
+	 * @memberOf SidebarAddedFragment
 	 * @type {!string}
 	 */
 	spritemap: Config.string().required(),
 };
 
-Soy.register(LayoutPageTemplateSidebarAddedFragment, templates);
+Soy.register(SidebarAddedFragment, templates);
 
-export {LayoutPageTemplateSidebarAddedFragment};
-export default LayoutPageTemplateSidebarAddedFragment;
+export {SidebarAddedFragment};
+export default SidebarAddedFragment;
