+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/sidebar/SidebarAddedFragment.soy
@@ -1,7 +1,7 @@
-{namespace LayoutPageTemplateSidebarAddedFragment}
+{namespace SidebarAddedFragment}
 
 /**
- * LayoutPageTemplateSidebarAddedFragment
+ * SidebarAddedFragment
  */
 {template .render}
 	{@param name: string}
