+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/sidebar/SidebarFragmentCollection.soy
@@ -1,7 +1,7 @@
-{namespace LayoutPageTemplateFragmentCollection}
+{namespace SidebarFragmentCollection}
 
 /**
- * LayoutPageTemplateFragmentCollection
+ * SidebarFragmentCollection
  */
 {template .render}
 	{@param fragmentCollection: [
